# openapi.yaml
openapi: 3.0.3
info:
  title: FIPE Data Ingestion API
  description: API para ingestão de dados do serviço FIPE e armazenamento em um banco de dados PostgreSQL.
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Servidor local

paths:
  /fipe/marcas:
    post:
      summary: Inicia a carga inicial das marcas
      responses:
        '200':
          description: Marcas geradas com sucesso
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: string
    get:
      summary: Busca todas as marcas armazenadas
      responses:
        '200':
          description: Lista de marcas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Marca'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: string

  /fipe/veiculos/{marcaCode}:
    get:
      summary: Busca todos os veículos de uma marca específica
      parameters:
        - name: marcaCode
          in: path
          required: true
          schema:
            type: integer
          description: ID da marca
      responses:
        '200':
          description: Lista de veículos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Veiculo'
        '404':
          description: Marca não encontrada
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: string

  /fipe/veiculos/{veiculoId}:
    put:
      summary: Atualiza os dados de um veículo específico
      parameters:
        - name: veiculoId
          in: path
          required: true
          schema:
            type: string
          description: ID do veículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Veiculo'
      responses:
        '200':
          description: Veículo atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
        '404':
          description: Veículo não encontrado
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: string

components:
  schemas:
    Marca:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        tipo:
          type: string
      required:
        - code
        - name
        - tipo

    Veiculo:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        tipo:
          type: string
        observation:
          type: string
      required:
        - code
        - name
        - tipo